src/auto-merge/requirements.txt
argparse
git-python
python-gitlab==2.10.1
atlassian-python-api
